# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/roles": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "查询数据",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "分页索引",
                        "name": "current",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "分页大小",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "查询值",
                        "name": "queryValue",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态(1:启用 2:禁用)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/schema.ListResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "list": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/schema.Role"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "创建数据",
                "parameters": [
                    {
                        "description": "创建数据",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.Role"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.IDResult"
                        }
                    },
                    "400": {
                        "description": "{error:{code:0,message:无效的请求参数}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            }
        },
        "/api/v1/roles.select": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "查询选择数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "查询值",
                        "name": "queryValue",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态(1:启用 2:禁用)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/schema.ListResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "list": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/schema.Role"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "{error:{code:0,message:未知的查询类型}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            }
        },
        "/api/v1/roles/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "查询指定数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "唯一标识",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Role"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "404": {
                        "description": "{error:{code:0,message:资源不存在}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "更新数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "唯一标识",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新数据",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.Role"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Role"
                        }
                    },
                    "400": {
                        "description": "{error:{code:0,message:无效的请求参数}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "删除数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "唯一标识",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status:OK}",
                        "schema": {
                            "$ref": "#/definitions/schema.StatusResult"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            }
        },
        "/api/v1/roles/{id}/disable": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "禁用数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "唯一标识",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status:OK}",
                        "schema": {
                            "$ref": "#/definitions/schema.StatusResult"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            }
        },
        "/api/v1/roles/{id}/enable": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "角色管理"
                ],
                "summary": "启用数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "唯一标识",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status:OK}",
                        "schema": {
                            "$ref": "#/definitions/schema.StatusResult"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schema.ErrorItem": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "type": "integer"
                },
                "message": {
                    "description": "错误信息",
                    "type": "string"
                }
            }
        },
        "schema.ErrorResult": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "错误项",
                    "$ref": "#/definitions/schema.ErrorItem"
                }
            }
        },
        "schema.IDResult": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "schema.ListResult": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "object"
                },
                "pagination": {
                    "$ref": "#/definitions/schema.PaginationResult"
                }
            }
        },
...
        "schema.PaginationResult": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "schema.Role": {
            "type": "object",
            "required": [
                "name",
                "role_menus",
                "status"
            ],
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "creator": {
                    "description": "创建者",
                    "type": "string"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "string"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "role_menus": {
                    "description": "角色菜单列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.RoleMenu"
                    }
                },
                "sequence": {
                    "description": "排序值",
                    "type": "integer"
                },
                "status": {
                    "description": "状态(1:启用 2:禁用)",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
        "schema.RoleMenu": {
            "type": "object",
            "required": [
                "action_id",
                "menu_id",
                "role_id"
            ],
            "properties": {
                "action_id": {
                    "description": "动作ID",
                    "type": "string"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "string"
                },
                "menu_id": {
                    "description": "菜单ID",
                    "type": "string"
                },
                "role_id": {
                    "description": "角色ID",
                    "type": "string"
                }
            }
        },
        "schema.StatusResult": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "状态(OK)",
                    "type": "string"
                }
            }
        },
...
    }
}